Unit tests: RPMImageInjectionStep
This step is the first from the unit-tested to use Routes and Builds, so
fake clientsets needed to be added for these resources. Testing the step
is straightforward and uses the established setup-examine-execute
pattern. This step depends on Builds having correct behavior (not just
Build resource having correct status at the end, but also creating the
Output, in this case an ImageStreamTag), so fake cluster behavior is
more complicated. This behavior is likely to be useful for testing more
steps, so it was extracted into a separate method in testing.go.